Address review comments
(cherry picked from commit cf19e539c8ddb344d1c6dcf3aa9b24d0f44f0e00)
